requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0

# Pydantic v2 и настройки
pydantic==2.5.0
pydantic-settings==2.1.0

# Аутентификация и безопасность
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Загрузка форм и email
python-multipart==0.0.6
email-validator==2.1.0
